Refactor DayCell and improve Calendar UI/UX
This commit refactors the `DayCell` into its own animated component, enhances the Calendar screen UI, and adds ProGuard rules for Apache POI.

- **`DayCell` Refactoring**:
    - Extracted the day item from `MonthDaysList` into a new, reusable `DayCell.kt` composable.
    - Added animations for selection (scale pop, elevation, and color changes) and press (scale shrink) to provide better visual feedback.

- **Calendar Screen UI/UX Improvements**:
    - The total hours displayed in the `MonthGrid` and `MonthlyTotalPill` now support `Double` values, allowing for more precise time representation (e.g., "7.5h").
    - Updated string resources and logic for sharing month summaries to handle decimal hour values.
    - Revamped the styling of `CategorySegments`, `MonthlyTotalPill`, and the `ShareMonthButton` for a more consistent and modern UI that aligns with the `MaterialTheme`.

- **Home Screen & Daily List**:
    - Introduced animations in `ListOfWorksDaily` to smoothly transition between the empty state (Lottie animation) and the list of work logs.
    - Minor padding adjustments in `HeaderWidget` and `MonthDaysList` for better alignment.

- **Build Configuration**:
    - Added extensive ProGuard rules to `proguard-rules.pro` to support the Apache POI library, preventing issues with minification in release builds.

+++ app/src/main/java/org/examples/time_manager/features/calendar/presentation/components/MonthOverviewModels.kt
@@ -7,7 +7,7 @@ import java.time.LocalDate
 @Immutable
 data class DayWork(
     val date: LocalDate,
-    val hours: Int,
+    val hours: Double,
 )
 
 @Immutable
